elon_agent/pyproject.toml
[project]
name = "elon-agent"
version = "0.1.0"
description = "Host coordinator agent for the A2A badminton scheduling system"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.11",
    "google-adk>=1.18.0",
    "httpx>=0.28.1",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.2.1",
]
